backend/requirements.txt
requests>=2.32.3
beautifulsoup4>=4.13.3
# pandas>=2.2.3
mysql-connector-python>=9.2.0
# sqlalchemy>=2.0.40
fastapi>=0.115.12